### 2024-09-13 #7196 @kyleniemeyer
Hi @thelfer, does your software package have online documentation? I saw the `docs/web` directory that it seems set up for that, but I can't find any links (like on the GitHub repo landing page).
