### 2024-12-11 #7196 @IgorBaratta
> @IgorBaratta @IvanYashchuk, would any of you have time and be willing to review this submission for JOSS? As you might know, we carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Thanks for your time!

Thank you for reaching out. Unfortunately, I am currently reviewing other papers and, as a result, won’t be able to take on this review at this time.
