+++ cmake/modules/mtest.cmake
@@ -90,6 +90,8 @@ function(add_mtest interface lib)
       set(_CMD _BEHAVIOUR)
     elseif(${_ARG} MATCHES LIBRARY)
       set(_CMD _LIBRARY)
+    elseif(${_ARG} MATCHES MATERIAL_PROPERTIES_LIBRARIES)
+      set(_CMD MATERIAL_PROPERTIES_LIBRARIES)
     elseif(${_ARG} MATCHES MATERIAL)
       set(_CMD _MATERIAL)
     elseif(${_ARG} MATCHES INTERFACE)
@@ -100,8 +102,6 @@ function(add_mtest interface lib)
       set(_CMD _ACCELERATION_ALGORITHM)
     elseif(${_ARG} MATCHES STIFFNESS_MATRIX_TYPE)
       set(_CMD _STIFFNESS_MATRIX_TYPE)
-    elseif(${_ARG} MATCHES MATERIAL_PROPERTIES_LIBRARIES)
-      set(_CMD MATERIAL_PROPERTIES_LIBRARIES)
     else()
       if(${_CMD} MATCHES TEST_NAME)
         list(APPEND _TESTS "${_ARG}")
