### 2024-12-20 7ab460e
fix madnex support
