+++ CMakeLists.txt
@@ -31,7 +31,7 @@ add_subdirectory(generic-behaviours)
 add_subdirectory(materials)
 add_subdirectory(mfront-book-examples)
 add_subdirectory(unit-tests/mfront-wrappers)
-# add_subdirectory(unit-tests/behaviours-without-temperature)
+add_subdirectory(unit-tests/behaviours-without-temperature)
 if(TFEL_MADNEX_SUPPORT)
 add_subdirectory(unit-tests/madnex/material-properties-1)
 add_subdirectory(unit-tests/madnex/material-properties-2)
