### 2025-01-09 #7196 @srmnitc
> @srmnitc Sorry for the delay.
>
> > Do you have any suggestions for reviewers?
>
> Are there a list of potentiel reviewers ?

This is the list: https://reviewers.joss.theoj.org/, but you can also suggest other people that might be relevant for the paper. I was travelling over the holidays. Now I am back, and will start the search for reviewers again. Thanks for being patient!
