### 2025-01-31 #7196 @jacobmerson
@srmnitc I'm generally willing to review this and have the relevant background, although I'm currently working on two other JOSS reviews (one from the same author as this). So, another reviewer may be able to turn it around faster.
