### 2025-02-02 #7196 @srmnitc
> [@srmnitc](https://github.com/srmnitc) I'm generally willing to review this and have the relevant background, although I'm currently working on two other JOSS reviews (one from the same author as this). So, another reviewer may be able to turn it around faster.

Thanks for letting us know!
