### 2025-02-02 #7742 @editorialbot
[REVIEW]: The MFrontGallery project
**Submitting author:** @thelfer (<a href="http://orcid.org/0000-0003-2460-5816">Thomas Helfer</a>)
**Repository:** https://github.com/thelfer/MFrontGallery
**Branch with paper.md** (empty if default branch): master
**Version:** 2.1
