### 2025-02-02 #7742 @srmnitc
@thelfer editorialbot is complaining that it could not find a license. This is probably because the license is not in a file named LICENSE. Is this something you could do?
