### 2025-02-02 #7742 @thelfer
@srmnitc I copied `LGPL-3.0.txt` in `LICENSE`
