### 2025-02-09 #7742 @mrbuche
> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

The paper is successful in explaining how this software fits in with compatible software, but perhaps not how it compares to equivalent software that may already be out there. Maybe there is no comparable alternatives, but that could then be stated. @srmnitc is this what is meant from this bullet?
