### 2025-02-10 #7742 @mrbuche
> > > **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
> >
> >
> > The paper is successful in explaining how this software fits in with compatible software, but perhaps not how it compares to equivalent software that may already be out there. Maybe there is no comparable alternatives, but that could then be stated. [@srmnitc](https://github.com/srmnitc) is this what is meant from this bullet?
>
> Yes, that point is meant to ensure that it is compared to other available tools. I agree with your assessment that if none exists, that could be stated.

Okay thank you! @thelfer that is my only criticism on the writing currently. I have gotten through most things except trying to install and use the software myself, things look great overall!
