### 2025-02-10 #7742 @thelfer
@mrbuche Thanks for this review. Concerning the state of the field, the question is somehow difficult because the project is really focused on QA issues related to material knowledge management which is seldom a major concern in most  open-source thermomechanical solvers.

Don't be mistaken by the last sentence. Open-source solvers are very concerned by quality issues regarding code, verifications tests, etc.. and do a wonderful job.

However, most (thermomechanical) solvers delegate material management issues to their users and for what I know, helping to set a proper material knowledge management strategy is not really dealt with in details.

The major benefit of the project (and the paper) is to highlight the need and to propose, thanks to `MFront`, a solution that can be used in a wide variety of context (languages, solvers, etc..)
