### 2025-02-11 #7742 @mrbuche
> [@mrbuche](https://github.com/mrbuche) Thanks for this review. Concerning the state of the field, the question is somehow difficult because the project is really focused on QA issues related to material knowledge management which is seldom a major concern in most open-source thermomechanical solvers.
>
> Don't be mistaken by the last sentence. Open-source solvers are very concerned by quality issues regarding code, verifications tests, etc.. and do a wonderful job.
>
> However, most (thermomechanical) solvers delegate material management issues to their users and for what I know, helping to set a proper material knowledge management strategy is not really dealt with in details.
>
> The major benefit of the project (and the paper) is to highlight the need and to propose, thanks to `MFront`, a solution that can be used in a wide variety of context (languages, solvers, etc..)

@thelfer that is definitely true I feel, the material management side of things is usually totally outside the software used to do numerical solutions, which is then usually put on the user to handle somehow, which is then usually something that doesn't have any sense of QA.

Maybe this is all technically clear in the paper, in places like "... the originality of the MFrontGallery project is to..." but just not the recipe that JOSS gives, which could be more like guidelines than requirements? Could point out that the software is "first of its kind" or something too. I am happy either way and am checking that box!
