### 2025-02-11 #7742 @thelfer
@mrbuche What about adding the following section at the end of the paper, after the conclusion and before the references ?

~~~~
# State of the field

The project is focused on Quality Assurance issues related to material
knowledge management which is, in the authors' experience, seldom a
major concern in most open-source thermomechanical solvers. Several
libraries providing high quality implementations of constitutive
equations are available, but generally dedicated to one specific solver.
The implementations are generally generic (as opposed to
self-contained). The `MFrontGallery` project thus provide a unique
approach.
~~~~
