### 2025-02-11 2c76e0b
Takes @mrbuche review into account (https://github.com/openjournals/joss-reviews/issues/7742)
