### 2025-02-17 #7742 @mrbuche
WIP

```sh
git clone git@github.com:thelfer/tfel.git
cd tfel/
mkdir build/
cd build/
cmake ../ -DCMAKE_BUILD_TYPE=Release -Dlocal-castem-header=ON -Denable-fortran=ON -Denable-aster=ON -DCMAKE_INSTALL_PREFIX=.
make -j 16
make install
cd ../../
export PATH=$PATH:/home/mrbuche/GitHub/tfel/build/lib/:/home/mrbuche/GitHub/tfel/build/bin/
git clone git@github.com:thelfer/MFrontGallery.git
cd MFrontGallery/
mkdir build/
cd build/
cmake ../
```
