### 2025-02-19 #7742 @thelfer
@mrbuche After installing `TFEL` from sources, you shall set some variable environments (`PATH`, `LD_LIBRARY_PATH`, `PYTHONPATH`). Generally, I use a script named `env.sh` that I place at the root of the installation directory. Here is a typicaly example:

```bash
export TFELHOME=/home/th202608/codes/tfel/master/install-python-3.10.12
export PATH=$TFELHOME/bin:$PATH
export LD_LIBRARY_PATH=$TFELHOME/lib:$LD_LIBRARY_PATH
export PYTHONPATH=$TFELHOME/lib/python3.10/site-packages/:$PYTHONPATH
```

Then I call `source <path_to_tfel>/env.sh` where you shall replace `<path_to_tfel>` by the path to the installation directory
