### 2025-02-19 #7742 @mrbuche
I was missing LD_LIBRARY_PATH originally and retried, editing above (except for the PYTHONPATH, I did not install TFEL with the Python interface option) and it still does not work for me, same error.
