### 2025-02-19 #7742 @thelfer
I followed your steps and it seems to work as expected:

~~~~
$ cmake ../ -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=.
-- The C compiler identification is GNU 11.4.0
-- The CXX compiler identification is GNU 11.4.0
-- Detecting C compiler ABI info
-- Detecting C compiler ABI info - done
-- Check for working C compiler: /usr/bin/cc - skipped
-- Detecting C compile features
-- Detecting C compile features - done
-- Detecting CXX compiler ABI info
-- Check for working CXX compiler: /usr/bin/c++ - skipped
-- Detecting CXX compile features
CMake Deprecation Warning at cmake/modules/mfm.cmake:1 (cmake_policy):
  The OLD behavior for policy CMP0005 will be removed from a future version
  of CMake.

  The cmake-policies(7) manual explains that the OLD behaviors of all
  policies are deprecated and that a policy should be set to OLD only under
  specific short-term circumstances.  Projects should be ported to the NEW
  behavior and not rely on setting a policy to OLD.
Call Stack (most recent call first):
  CMakeLists.txt:16 (include)

CMake Deprecation Warning at cmake/modules/mfm.cmake:2 (cmake_policy):
  The OLD behavior for policy CMP0026 will be removed from a future version
  of CMake.

  The cmake-policies(7) manual explains that the OLD behaviors of all
  policies are deprecated and that a policy should be set to OLD only under
  specific short-term circumstances.  Projects should be ported to the NEW
  behavior and not rely on setting a policy to OLD.
Call Stack (most recent call first):
  CMakeLists.txt:16 (include)

unknown option --python-version
Valid options are:
--all             : request flags for all librairies.
--castem          : request flags for castem.
--compiler-flags  : return TFEL's recommended compiler flags.
--config          : request flags for TFELConfig.
--cppflags        : return preprocessor flags.
--cxx-standard    : print the version of the C++ standard used to compile TFEL.
--debug-flags     : return flags adding debugging symbols.
--exceptions      : request flags for TFELException.
--glossary        : request flags for TFELGlossary.
--help            : print this help message.
--include-path    : return the path to the `TFEL` headers.
--includes        : return tfel include path.
--ldflags         : return linking flags.
--library-dependency: return the dependencies of a `TFEL` library.
--library-path    : return the path to the `TFEL` library.
--libs            : return linking flags.
--licence         : print TFEL licences.
--madnex-support  : print if mandex support is enabled (display `true` or `false`).
--major-number    : print tfel major version.
--material        : request flags for TFELMaterial.
--math            : request flags for TFELMath.
--math-cubic-spline: request flags for TFELMathCubicSpline.
--math-kriging    : request flags for TFELMathKriging.
--math-parser     : request flags for TFELMathParser.
--mfront          : request flags for TFELMFront.
--mfront-doc-has-standalone-option: print if mfront-doc has the standalone option
--mfront-log-stream: request flags for MFrontLogStream.
--mfront-profiling: request flags for MFrontProfiling.
--minor-number    : print tfel minor version.
--mtest           : request flags for TFELMTest.
--numodis         : request flags for TFELNUMODIS.
--oflags          : return tfel recommended optimisation flags with architecture specific flags.
--oflags0         : return tfel recommended optimisation flags without architecture specific flags.
--oflags2         : return some aggressive optimisation flags, possibly at the expense of numerical precision. This shall be added to `--oflags` results.
--python-bindings-support: print the python binding are supported (display `true` or `false`).
--quiet-failure   : quietly fails without error message
--revision-number : print tfel revision version.
--system          : request flags for TFELSystem.
--tests           : request flags for TFELTests.
--tfel-home       : returns the environment variable in which the TFEL's installation path shall be defined
--unicode-support : request flags for TFELUnicodeSupport.
--utilities       : request flags for TFELUtilities.
--version         : print tfel version and svn revision.
--warning         : return tfel recommended warnings.
-- no tfel python bindings
-- tfel version          : 5.1.0-dev
-- tfel C++ standard     : 20
-- mfront                : /tmp/mrbuche/build/tfel/build/bin/mfront
-- mfront-doc            : /tmp/mrbuche/build/tfel/build/bin/mfront-doc
-- mfront-query          : /tmp/mrbuche/build/tfel/build/bin/mfront-query
-- tfel-config           : /tmp/mrbuche/build/tfel/build/bin/tfel-config
-- tfel-check            : /tmp/mrbuche/build/tfel/build/bin/tfel-check
-- tfel include          : /tmp/mrbuche/build/tfel/build/include
-- tfel libs             : /tmp/mrbuche/build/tfel/build/lib
-- TFELTests             : /tmp/mrbuche/build/tfel/build/lib/libTFELTests.so
-- TFELTests             : /tmp/mrbuche/build/tfel/build/lib/libTFELTests.so
-- TFELException         : /tmp/mrbuche/build/tfel/build/lib/libTFELException.so
-- TFELUtilities         : /tmp/mrbuche/build/tfel/build/lib/libTFELUtilities.so
-- TFELMath              : /tmp/mrbuche/build/tfel/build/lib/libTFELMath.so
-- TFELMaterial          : /tmp/mrbuche/build/tfel/build/lib/libTFELMaterial.so
-- TFELPhysicalConstants :
-- madnex support disabled
-- fortran support disabled
-- enabling flag 'Wno-conversion'
-- enabling flag 'Wall'
-- enabling flag 'W'
-- enabling flag 'Wshadow'
-- enabling flag 'Wpointer-arith'
-- enabling flag 'Wcast-qual'
-- enabling flag 'Wcast-align'
-- enabling flag 'Wwrite-strings'
-- enabling flag 'Wctor-dtor-privacy'
-- enabling flag 'Wnon-virtual-dtor'
-- enabling flag 'Wno-endif-labels'
-- enabling flag 'Wmissing-format-attribute'
-- enabling flag 'Wno-multichar'
-- enabling flag 'Wno-deprecated-declarations'
-- enabling flag 'Wpacked'
-- enabling flag 'Wredundant-decls'
-- enabling flag 'Wunknown-pragmas'
-- enabling flag 'Wreorder'
-- enabling flag 'Wno-conversion'
-- enabling flag 'Winit-self'
-- enabling flag 'Wlogical-op'
-- enabling flag 'Wmissing-include-dirs'
-- enabling flag 'Wnoexcept'
-- enabling flag 'Wstrict-null-sentinel'
-- enabling flag 'Wtrampolines'
-- enabling flag 'Wvector-operation-performance'
-- enabling flag 'Wignored-qualifiers'
-- enabling flag 'Wsequence-point'
-- enabling flag 'Wswitch'
-- enabling flag 'Wsuggest-final-types'
-- enabling flag 'Wsuggest-final-methods'
-- enabling flag 'Wsuggest-override'
-- enabling flag 'Wswitch-bool'
-- enabling flag 'Wlogical-not-parentheses'
-- enabling flag 'Wsizeof-array-argument'
-- enabling flag 'Wbool-compare'
-- enabling flag 'Wshift-negative-value'
-- enabling flag 'Wshift-overflow'
-- enabling flag 'Wtautological-compare'
-- enabling flag 'Wnull-dereference'
-- enabling flag 'Wduplicated-cond'
-- enabling flag 'Wmisleading-indentation'
-- enabling flag 'Wduplicated-branches'
-- enabling flag 'Wregister'
-- enabling flag 'Wint-in-bool-context'
-- enabling flag 'Wsuggest-override'
-- enabling flag 'Wno-sign-compare'
-- enabling flag 'Wno-missing-declarations'
-- flag '/Wno-weak-vtables' disabled
-- enabling flag 'Wdisabled-optimization'
-- enabling flag 'fvisibility=hidden'
-- enabling flag 'fvisibility-inlines-hidden'
-- enabling flag 'march=native'
-- enabling flag 'ftree-vectorize'
-- enabling flag 'fno-fast-math'
-- enabling flag 'ffast-math'
-- enabling [[ noreturn ]] attribute
-- Enabling generation of documentation based on pandoc
-- pandoc:          /home/th202608/.cabal/bin/pandoc
-- pandoc-crossref: /home/th202608/.cabal/bin/pandoc-crossref
-- Found Gnuplot: /usr/bin/gnuplot (found version "5.4.2")
-- Configuring done
-- Generating done
-- Build files have been written to: /tmp/mrbuche/build/MFrontGallery/build
~~~~

Note that nothing will be done since you did not select any target. I did the same with `-Denable-aster` and it worked fine.
