### 2025-02-22 #7742 @likask
Declaimer. We integrated MFront into our code, and we are using Spack for installation, which goes straightforward in several systems and cases which we have tried.
