### 2025-02-28 #7742 @thelfer
> [@thelfer](https://github.com/thelfer) [@srmnitc](https://github.com/srmnitc) alright I got it working in a Docker container, everything builds and installs as described, running and testing it as well. I have finished my checklist and everything looks great!

Great !
