+++ docs/web/SemiExplicitModifiedCamClay_OpenGeoSys2023.md
@@ -1,8 +1,8 @@
-# Introduction {#sec:introduction}
+# Introduction $^1$
 
 The Cambridge (Cam) clay model
 [@Borja1990; @Borja1991; @Borja1998; @Callari1998; @Peric2006] describes
-the stress-dependent deformation behaviour of cohesive soils.[^2]
+the stress-dependent deformation behaviour of cohesive soils.$^2$
 Thereby, effects like
 
 1.  elasto-plastic deformation,
@@ -28,13 +28,13 @@ of underscores indicates the order of the tensor object.
 ## Preliminaries
 
 In the small-strain setting, there is an additive split of the linear
-strain tensor reading $$\label{eq:elasticplastic}
+strain tensor reading $$
   \underline{\varepsilon }= \underline{\varepsilon}_\text{e}+ \underline{\varepsilon}_\text{p}\ .$$
 The generalized Hooke's law relates elastic strains with stresses as
-$$\label{eq:Hooke}
+$$
   \underline{\sigma }= \underline{\underline{D}} \colon\underline{\varepsilon}_\text{e}\ .$$
-Splitting the stress tensor[^3] with respect to deviatoric and
-volumetric parts yields $$\label{eq:decomposition}
+Splitting the stress tensor $^3$ with respect to deviatoric and
+volumetric parts yields $$
   \underline{\sigma }= \underline{\sigma }^\text{D}+ \frac{1}{3}I_1(\underline{\sigma})\underline{I} \ .$$
 Therewith, the von-Mises stress and the hydrostatic pressure is defined
 as
@@ -48,11 +48,14 @@ $$\frac{\partial q}{\partial \underline{\sigma}} = \frac{3}{2}\ \frac{\underline
 Dealing with porous media there is a kinematic relation between porosity
 and volumetric strain. Let the total volume of some representative
 elementary volum (REV) be divided into pore volume and solid volume:
-$$V = V_{\text{S}} + V_{\text{P}} \ .$$ The porosity is defined as the
+$$V = V_\text{S} + V_\text{P}\ .$$ 
+The porosity is defined as the
 pore volume fraction, i. e. $\phi=V_{\text{P}}/V$. Evaluating the mass
 balance of the porous medium (incompressible solid phase) yields the
-porosity evolution $$\label{eq:porEvolution}
+porosity evolution
+$$
   \dot{\phi} - \phi\mathrm{div}(\dot{\vec u}) = \mathrm{tr}(\dot{\underline{\varepsilon}})\ .$$
+
 Exploiting
 $\mathrm{div}(\dot{\vec u})\equiv \mathrm{tr}(\dot{\underline{\varepsilon}})=\dot{\varepsilon}^\text{V}$
 and separating the variables, this differential equation can be solved
@@ -63,67 +66,75 @@ the porosity (evolution) can be calculated from
 $\varepsilon^\text{V}_\text{p}$ only. Instead of the porosity $\phi$,
 the *pore number* $e=V_{\text{P}} / V_{\text{S}}$ (a. k. a. void ratio)
 or the *volume ratio* $v=V / V_{\text{S}}$ can equally be used with the
-relations $$\label{eq:e-phiRelation}
-  e = \frac{\phi}{1-\phi} \quad\text{with}\quad v = 1+e = (1-\phi)^{-1}\ .$$
-From [\[eq:porEvolution\]](#eq:porEvolution){reference-type="eqref"
-reference="eq:porEvolution"} and
-[\[eq:e-phiRelation\]](#eq:e-phiRelation){reference-type="eqref"
[…]
