### 2025-04-22 0e98bf8
remove support for europlexus interface (it is deprecated in favour of the generic interface)
