+++ .github/workflows/cmake.yml
@@ -40,7 +40,7 @@ jobs:
               -Denable-octave=OFF -Denable-generic=ON -Denable-castem=ON \
               -Denable-castem-behaviours=ON -Denable-aster=ON            \
               -Denable-cyrano=ON -Denable-ansys=ON                       \
-              -Denable-europlexus=ON -Denable-calculix=ON                \
+              -Denable-calculix=ON                                       \
               -Denable-abaqus=ON -Denable-diana-fea=ON                   \
               -Denable-zmat=OFF -Denable-fortran-behaviours-wrappers=ON  \
               -Denable-mfront-documentation-generation=OFF
@@ -54,11 +54,3 @@ jobs:
       run: | 
         spack load tfel
         make check
-        
-    # - name: Test Report
-    #   uses: dorny/test-reporter@v1
-    #   if: success() || failure()    # run this step even if previous step failed
-    #   with:
-    #     name: Tests reports         # Name of the check run which will be created
-    #     path: build/**/*.xml        # Path to test results
-    #     reporter: jest-junit
