### 2025-04-22 303e778
Some tests requires the generic-models to be selected
