### 2025-04-23 #7742 @thelfer
@srmnitc

> Double check authors and affiliations (including ORCIDs)

Done

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Version 2.1 is fine, there was no major changes due to the review

> Archive the release on Zenodo/figshare/etc and post the DOI here.

Done: https://zenodo.org/records/13258209

> Make sure that the license listed for the archive is the same as the software licen

Yes

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Well, I published the Zenodo archive as the project manager, I did not pay attention to the fact that all the contributors shall be mentioned. Is that required ? If so, shall I publish a new archive ?
