### 2025-05-19 thelfer/MFrontGallery#63 @srmnitc
Changes to the JOSS paper
This is part of review over at https://github.com/openjournals/joss-reviews/issues/7742

Some small changes to the paper:
- Affiliation 5: Please make sure that the address is complete
- line 25: could you please add linkes for MANTA and XPER as well?
- State of the field - comes after the Acknowledgements. Would it be possible to move this up, or merge into other sections?
