### 2025-05-19 6b0d70f
Fix T. Nagel' affiliation, add references for Manta and XPER
