+++ docs/papers/joss/bibliography.bib
@@ -1,3 +1,30 @@
+@unpublished{perales_xper_2022,
+	title = {{XPER} : une plateforme pour la simulation numérique distribuée d'interactions multiphysiques entre corps.},
+	url = {https://hal.science/hal-03704427},
+	shorttitle = {{XPER}},
+	abstract = {{XPER} est un outil numérique parallélisé dédié aux interactions multiphysiques entre corps déformables. Les interactions concernent la rupture dynamique de matériaux hétérogènes dans le cadre du couplage thermo-chimio-poromécanique. Le logiciel repose sur le couplage du logiciel {LMGC}90
+(Dynamique des Contacts) pour la prise en compte d’interactions complexes entre les corps et de la bibliothèque {PELICANS} pour la résolution des comportements volumiques Eléments Finis des corps. {XPER} est développé en commun par l’{IRSN}/{CNRS}/Université de Montpellier (laboratoire {MIST}).},
+	author = {Perales, Frederic and Socie, Adrien and Nkoumbou Kaptchouang, Noé Brice and Dubois, Frederic and Monerie, Yann and Mozul, Remy and Vincent, Pierre Guy and Babik, Fabrice},
+	urldate = {2023-11-20},
+	date = {2022},
+	note = {Published: {CSMA} 2022 15ème Colloque National en Calcul des Structures, {CSMA} 2022},
+	file = {HAL PDF Full Text:/home/th202608/.zotero/zotero/vzpzva46.default/zotero/storage/K7RRAQ7K/Perales et al. - 2022 - XPER  une plateforme pour la simulation numérique.pdf:application/pdf},
+}
+
+@inproceedings{jamond_manta_2024,
+	location = {Giens, France},
+	title = {{MANTA}: an industrial-strength open-source high performance explicit and implicit multi-physics solver},
+	url = {https://hal.science/hal-04610968},
+	shorttitle = {{MANTA}},
+	booktitle = {16ème Colloque National en Calcul de Structures},
+	publisher = {{CNRS}, {CSMA}, {ENS} Paris-Saclay, {CentraleSupélec}},
+	author = {Jamond, Olivier and Lelong, Nicolas and Brooking, Guillaume and Helfer, Thomas and Prabel, Benoit and Prat, Raphael and Jaccon, Adrien},
+	urldate = {2024-07-29},
+	date = {2024-05},
+	keywords = {{HPC}, fluid mechanics, Industrial applications, multiphysics coupling, {PDEs} solving, structural mechanics},
+	file = {HAL Snapshot:/home/th202608/.zotero/zotero/vzpzva46.default/zotero/storage/U8WSDGNR/hal-04610968.html:text/html},
+}
+
 @article{bernaud_pleiades_2024,
 	title = {{PLEIADES}: A numerical framework dedicated to the multiphysics and multiscale nuclear fuel behavior simulation},
 	volume = {205},
