### 2025-05-19 25887f8
Fix T. Nagel' affiliation (my bad, they were right)
