### 2025-05-19 thelfer/MFrontGallery#63 @thelfer
Great, I am waiting for Lorenzo's confirmation. Thanks
