### 2025-05-19 thelfer/MFrontGallery#63 @thelfer
Lorenzo confirms that the affiliation and address are fine, so everything is ok on our side.
