### 2025-05-23 #7742 @kyleniemeyer
@thelfer in the Zenodo archive, can you clean up the author names? I think your first name is missing, and a coauthor's surname is capitalized unnecessarily.

In the paper, please remove the formatting from MFrontGallery in the paper title—it will cause markup code to appear in the published page title.

Also, the paper references section numbers a few times, but JOSS papers do not have numbered sections. Can you please remove those numbers? (I'd actually suggest just removing intra-paper references altogether, since the paper is short.)
