### 2025-05-23 cad2f96
remove the formatting from MFrontGallery in the paper title
