### 2025-05-23 #7742 @thelfer
@kyleniemeyer

> @thelfer in the Zenodo archive, can you clean up the author names? I think your first name is missing, and a coauthor's surname is capitalized unnecessarily.

Done

> please remove the formatting from MFrontGallery in the paper title

Done
