### 2025-05-23 #7742 @thelfer
@kyleniemeyer I did not take into account your remarks about references to sections. This is now done.
